pgbackrest: fix schedules (tpo/tpa/team#42043)
The problem with the previous "monthly" and "weekly" configurations is
that they were splaying the results from *after* the calendar
date. This *eventually* works, it converges over the planned
frequency, but when systemd restarts or the server reboots, things go
out of whack and there will be additional delays.

I consider this to be a bug in systemd, see:

https://github.com/systemd/systemd/issues/21166

So instead we pick a random day of the month or of the week in
Puppet. We also shift the definition to the client-side, because
otherwise the randomization happens on the server, and only on the
server, which defeats the purpose as all instances of the template
will run at the same time.