retry backups on failure
We had a situation where a backup had failed, presumably because it
was performed while the host was rebooting. It seems fair to try to
reboot those correctly.

We introduce exponential backoff here, but that's supported only in
systemd 254 (trixie and later). It's ignored by earlier versions
though, so that's safe to include here.

We limit the number of restarts per day to 10 as well.